Limited viewing access of assignments.
Create separate tables for all, tutor, and professor assignment lists only viewable if the user has that role.

Limited access to view and edit pages if the user has is associated with it.

Limited access to add page to only admins.

TODO: Only admins should be able to access the edit page.
  Redirecting with improper credentials should bring the user to an Access Denied page.