Revamped report pages.
Updated database to have a session date instead of submit date.

Updated the controller to convert the date into a DateTime object to store.

Updated the controller to convert the times into a DateTime object to store.

Added a datalist linked to the start and end times to show the proper format of time.

Updated all display pages to show dates and times in proper format.